+++ app/services/parts/create_part.rb
@@ -14,6 +14,7 @@ module Parts
     def call
       part = @user.parts.new(@params)
       if part.save
+        @user.grant_swap_meet_post_rewards!(part_count: @user.parts.count)
         Result.new(success?: true, part: part)
       else
         Result.new(success?: false, part: part)